调整操作用户登录验证

+++ WinWarBusiness/NewsBusiness.cs
@@ -80,7 +80,7 @@ namespace WinWarBusiness
         /// <param name="pageSize">每页新闻数</param>
         /// <param name="newsCode">最大新闻Code,第一页传 0</param>
         /// <returns></returns>
-        public List<NewsEntity> GetNews(string keyWords, int typeid, int pageSize, int userid, ref long newsCode)
+        public List<NewsEntity> GetNews(string keyWords, int typeid, int pageSize, long userid, ref long newsCode)
         {
             List<NewsEntity> list = new List<NewsEntity>();
 
@@ -134,7 +134,7 @@ namespace WinWarBusiness
         /// <param name="pageIndex"></param>
         /// <param name="userid"></param>
         /// <returns></returns>
-        public List<NewsCommentEntity> GetNewsComments(long newsCode, int pageSize, int userid, ref long id)
+        public List<NewsCommentEntity> GetNewsComments(long newsCode, int pageSize, long userid, ref long id)
         {
             List<NewsCommentEntity> list = new List<NewsCommentEntity>();
             DataTable dt = NewsDAL.BaseDAL.GetNewsComments(newsCode, pageSize, userid, ref id);
@@ -169,7 +169,7 @@ namespace WinWarBusiness
         /// <param name="isAdd">true 点赞；false 取消点赞</param>
         /// <param name="userid">用户ID</param>
         /// <returns></returns>
-        public bool AddNewsPraiseCount(long newsCode, bool isAdd, int userid)
+        public bool AddNewsPraiseCount(long newsCode, bool isAdd, long userid)
         {
             return NewsDAL.BaseDAL.AddNewsPraiseCount(newsCode, isAdd, userid);
         }
@@ -181,7 +181,7 @@ namespace WinWarBusiness
         /// <param name="isAdd">true 点赞；false 取消点赞</param>
         /// <param name="userid">用户ID</param>
         /// <returns></returns>
-        public bool AddNewsCollectCount(long newsCode, bool isAdd, int userid)
+        public bool AddNewsCollectCount(long newsCode, bool isAdd, long userid)
         {
             return NewsDAL.BaseDAL.AddNewsCollectCount(newsCode, isAdd, userid);
         }
@@ -197,7 +197,7 @@ namespace WinWarBusiness
         /// <param name="replyUserID"></param>
         /// <param name="replyUserName"></param>
         /// <returns></returns>
-        public bool AddNewsComment(string content, long newsCode, int userid, string userName, long replyid, long replyUserID, string replyUserName)
+        public bool AddNewsComment(string content, long newsCode, long userid, string userName, long replyid, long replyUserID, string replyUserName)
         {
             return NewsDAL.BaseDAL.AddNewsComment(content, newsCode, userid,userName,replyid,replyUserID,replyUserName);
         }
